Removing more boilerplate code
The React bootstrapping script, create-react-app, does a whole lot more
than just create a simple HTML page and JS file for React. It actually
also does things like setting up a file watcher for auto-updates. In
the interests of keeping this app simple, I'm removing
registerServiceWorker.js (the file watcher). It's great for developing
with React, but not so helpful for *teaching* it to a reader of this
source code.

Also greatly simplifying the actual HTML file.